pyproject.toml
[project]
name = "cafe_finder"
version = "0.1.0"
description = "This is an interactive program to help find cute cafes near you!"
authors = [
    {name = "t3mi",email = "temi.akinmola24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    # provider-specific packages removed; add the API client you prefer (e.g. yelpapi) with `poetry add`
    "geopy (>=2.4.1,<3.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "haversine (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
